@@ -1,73 +1,81 @@
 """Util functions."""
 
 import os
+import typing as t
+from collections.abc import Generator
 from enum import Enum
-from typing import Any, Generator, Union
 
 from json_arrays import _types
 
-# pylint: disable=unsubscriptable-object
 
-
-def get_name_of_file(fp: _types.File) -> str:
-    return fp.name if hasattr(fp, "name") else ""
+def get_name_of_file(fp: _types.File | None) -> str:
+    """Return the name if fp has the attribute 'name' otherwise empty string."""
+    return str(fp.name) if fp and hasattr(fp, "name") else ""
 
 
 def is_jsonl(name: _types.Pathlike) -> bool:
     """Test if a filename is json lines.
+
     Also returns True for '<stdin>' and '<stdout>'.
 
     Arguments:
-        name {str} -- name to test
+        name: name to test
 
     Returns:
-        {bool} -- True if this name is jsonl.
+        True if this name is jsonl.
     """
-    if name in ["<stdin>", "<stdout>"]:
+    if name in {"<stdin>", "<stdout>"}:
         return True
-    base, suffix = os.path.splitext(str(name))
-    if suffix in [".gz", ".bz2"]:
-        _, suffix = os.path.splitext(base)
+    base, suffix = os.path.splitext(str(name))  # noqa: PTH122
+    if suffix in {".gz", ".bz2"}:
+        _, suffix = os.path.splitext(base)  # noqa: PTH122
     # print('suffix = {suffix}'.format(suffix=suffix))
-    return suffix in [".jsonl", ".jl", ".ndjson"]
+    return suffix in {".jsonl", ".jl", ".ndjson"}
 
 
 def is_gzip(name: _types.Pathlike) -> bool:
     """Test if a filename is gzip."""
-    _, suffix = os.path.splitext(str(name))
+    _, suffix = os.path.splitext(str(name))  # noqa: PTH122
     return suffix == ".gz"
 
 
 def is_bzip2(name: _types.Pathlike) -> bool:
-    """Test if a filename is gzip."""
-    _, suffix = os.path.splitext(str(name))
+    """Test if a filename is bzip2."""
+    _, suffix = os.path.splitext(str(name))  # noqa: PTH122
     return suffix == ".bz2"
 
 
-def to_bytes(s: Union[str, bytes, bytearray]) -> Union[bytes, bytearray]:
-    """Convert str to bytes, otherwise pass through s
+def to_bytes(s: str | bytes | bytearray) -> bytes | bytearray:
+    """Convert str to bytes, otherwise pass through.
 
     Args:
-        s (Union[str, bytes, bytearray]): the argument to assure bytes
+        s: the argument to assure bytes
 
     Returns:
-        Union[bytes, bytearray]: the possible converted argument
+         the possible converted argument
     """
-    return s if isinstance(s, (bytes, bytearray)) else s.encode("utf-8")
+    return s if isinstance(s, bytes | bytearray) else s.encode("utf-8")
 
 
 class Sink:
-    def __init__(self, sink: Generator[None, Any, None]):
+    """A utility that wraps a co-routine, and closes it."""
+
+    def __init__(self, sink: Generator[None, t.Any, None]) -> None:
+        """Wrap the given sink."""
         self.sink = sink
 
-    def __enter__(self):
+    def __enter__(self) -> Generator[None, t.Any, None]:
+        """Prepare the generator and return it."""
         next(self.sink)
         return self.sink
 
-    def __exit__(self, exc_type, exc_value, exc_traceback):
+    def __exit__(self, exc_type: t.Any, exc_value: t.Any, exc_traceback: t.Any) -> None:
+        """Call close on the wrapped generator."""
         self.sink.close()
 
 
 class JsonFormat(str, Enum):
+    """Format of json."""
+
     JSON = "json"
     JSON_LINES = "jsonl"
