@@ -1,10 +1,12 @@
+"""Example to show how to count a length of a json/ndjson file."""
+
 import sys
 
 import json_arrays
 
 
-def main():
-    """Count number of elements in a array.
+def main() -> None:
+    r"""Count number of elements in a array.
 
     Reads either the given path or from `stdin`.
 
@@ -13,7 +15,7 @@ def main():
     Should print 2
     """
     path = sys.argv[1] if len(sys.argv) > 1 else None
-    length = sum((1 for _ in json_arrays.load_from_file(path, use_stdin_as_default=True)))
+    length = sum(1 for _ in json_arrays.load_from_file(path, use_stdin_as_default=True))
     print(length)
 
 
