@@ -1,7 +1,5 @@
 """Exceptions for json_arrays."""
 
 
-class JsonStreamsException(Exception):
-    """
-    Base exception for json_arrays
-    """
+class JsonStreamsError(Exception):
+    """Base exception for json_arrays."""
