@@ -1,5 +1,6 @@
 import pytest
 
+from json_arrays import _types
 from json_arrays.utility import get_name_of_file, is_bzip2, is_gzip, is_jsonl
 
 
@@ -17,7 +18,7 @@ from json_arrays.utility import get_name_of_file, is_bzip2, is_gzip, is_jsonl
         "test.ndjson.bz2",
     ],
 )
-def test_filename_is_jsonl(filename: str):
+def test_filename_is_jsonl(filename: str) -> None:
     assert is_jsonl(filename)
 
 
@@ -25,7 +26,7 @@ def test_filename_is_jsonl(filename: str):
     "filename",
     ["test.json.gz", "test.jsonl.gz", "test.jl.gz", "test.ndjson.gz"],
 )
-def test_filename_is_gzip(filename: str):
+def test_filename_is_gzip(filename: str) -> None:
     assert is_gzip(filename)
 
 
@@ -33,7 +34,7 @@ def test_filename_is_gzip(filename: str):
     "filename",
     ["test.json.bz2", "test.jsonl.bz2", "test.jl.bz2", "test.ndjson.bz2"],
 )
-def test_filename_is_bzip2(filename: str):
+def test_filename_is_bzip2(filename: str) -> None:
     assert is_bzip2(filename)
 
 
@@ -44,7 +45,7 @@ def test_filename_is_bzip2(filename: str):
         "test.json.gz",
     ],
 )
-def test_filename_is_not_jsonl(filename: str):
+def test_filename_is_not_jsonl(filename: str) -> None:
     assert not is_jsonl(filename)
 
 
@@ -52,7 +53,7 @@ def test_filename_is_not_jsonl(filename: str):
     "filename",
     ["test.json", "test.jsonl", "test.jl", "test.ndjson"],
 )
-def test_filename_is_not_gzip(filename: str):
+def test_filename_is_not_gzip(filename: str) -> None:
     assert not is_gzip(filename)
 
 
@@ -60,10 +61,10 @@ def test_filename_is_not_gzip(filename: str):
     "filename",
     ["test.json", "test.jsonl", "test.jl", "test.ndjson"],
 )
-def test_filename_is_not_bzip2(filename: str):
+def test_filename_is_not_bzip2(filename: str) -> None:
     assert not is_bzip2(filename)
 
 
 @pytest.mark.parametrize("fp, expected", [(None, "")])
-def test_get_name_of_file(fp, expected):
-    assert get_name_of_file(fp) == expected
+def test_get_name_of_file(fp: _types.File | None, expected: str) -> None:
+    assert get_name_of_file(fp) == expected  # type: ignore[type-var]
